+++ packages/dart/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.1.13 (2026--)
 
 - **BREAKING CHANGES**: `strToNumberHash` has renamed to `numberHash`
+- `md5Hash`, `sha1Hash`, `sha256Hash`: Add an encoding option for hash functions
 - Add `sortNumeric` method
 
 ## 1.1.12 (2026-03-31)